Move uniqueIssue to shared
