+++ src/applications/appeals/995/components/AddIssue.jsx
@@ -10,11 +10,7 @@ import { $ } from 'platform/forms-system/src/js/utilities/ui';
 import recordEvent from 'platform/monitoring/record-event';
 
 import { checkValidations } from '../validations';
-import {
-  uniqueIssue,
-  missingIssueName,
-  maxNameLength,
-} from '../validations/issues';
+import { missingIssueName, maxNameLength } from '../validations/issues';
 import { validateDate } from '../validations/date';
 import { content } from '../content/addIssue';
 
@@ -26,6 +22,7 @@ import {
 } from '../../shared/constants';
 import { calculateIndexOffset, getSelected } from '../../shared/utils/issues';
 import { setStorage } from '../../shared/utils/addIssue';
+import { uniqueIssue } from '../../shared/validations/issues';
 
 const ISSUES_PAGE = `/${CONTESTABLE_ISSUES_PATH}`;
 const REVIEW_AND_SUBMIT = '/review-and-submit';
