+++ src/applications/appeals/995/tests/validations/issues.unit.spec.js
@@ -4,7 +4,6 @@ import sinon from 'sinon';
 import { getDate } from '../../utils/dates';
 import {
   checkIssues,
-  uniqueIssue,
   maxIssues,
   missingIssueName,
   maxNameLength,
@@ -122,58 +121,6 @@ describe('checkIssues', () => {
   it('should show an error with invalid selected additional issues date', () => {
     const errors = { addError: sinon.spy() };
     checkIssues(errors, _, _, _, _, _, getData({ aiDate: '2021-?-?' }));
-    expect(errors.addError.called).to.be.true;
-  });
-});
-
-describe('uniqueIssue', () => {
-  const _ = null;
-  const contestedIssues = [
-    {
-      attributes: {
-        ratingIssueSubjectText: 'test',
-        approxDecisionDate: '2021-01-01',
-      },
-    },
-  ];
-
-  it('should not show an error when there are no issues', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {});
-    expect(errors.addError.notCalled).to.be.true;
-  });
-  it('should not show an error when there are duplicate contested issues', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues: [contestedIssues[0], contestedIssues[0]],
-    });
-    expect(errors.addError.notCalled).to.be.true;
-  });
-  it('should not show an error when there are no duplicate issues (only date differs)', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues,
-      additionalIssues: [{ issue: 'test', decisionDate: '2021-01-02' }],
-    });
-    expect(errors.addError.notCalled).to.be.true;
-  });
-  it('should show an error when there is a duplicate additional issue', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues,
-      additionalIssues: [{ issue: 'test', decisionDate: '2021-01-01' }],
-    });
-    expect(errors.addError.called).to.be.true;
-  });
-  it('should show an error when there are multiple duplicate additional issue', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues,
-      additionalIssues: [
-        { issue: 'test2', decisionDate: '2021-02-01' },
-        { issue: 'test2', decisionDate: '2021-02-01' },
-      ],
-    });
     expect(errors.addError.called).to.be.true;
   });
 });
