+++ src/applications/appeals/995/validations/issues.js
@@ -8,20 +8,6 @@ import {
   getIssueName,
   getSelected,
 } from '../../shared/utils/issues';
-import { addUniqueIssueErrorMessage } from '../../shared/validations/issues';
-
-// Alert Veteran to duplicates based on name & decision date
-export const uniqueIssue = (
-  errors,
-  _fieldData,
-  formData,
-  _schema,
-  _uiSchema,
-  _index,
-  appStateData,
-) => {
-  addUniqueIssueErrorMessage(errors, formData, appStateData, errorMessages);
-};
 
 export const maxIssues = (errors, data) => {
   if (getSelected(data).length > MAX_LENGTH.SELECTIONS) {
