+++ src/applications/appeals/996/components/AddIssue.jsx
@@ -13,11 +13,7 @@ import recordEvent from 'platform/monitoring/record-event';
 // https://github.com/department-of-veterans-affairs/va.gov-team/issues/33797
 import { setData } from 'platform/forms-system/src/js/actions';
 
-import {
-  uniqueIssue,
-  missingIssueName,
-  maxNameLength,
-} from '../validations/issues';
+import { missingIssueName, maxNameLength } from '../validations/issues';
 import { validateDate } from '../validations/date';
 import { content } from '../content/addIssue';
 
@@ -28,8 +24,9 @@ import {
   SELECTED,
 } from '../../shared/constants';
 import { calculateIndexOffset, getSelected } from '../../shared/utils/issues';
-import { checkValidations } from '../../shared/validations/issues';
+import { checkValidations } from '../../shared/validations';
 import { setStorage } from '../../shared/utils/addIssue';
+import { uniqueIssue } from '../../shared/validations/issues';
 
 const ISSUES_PAGE = `/${CONTESTABLE_ISSUES_PATH}`;
 const REVIEW_AND_SUBMIT = '/review-and-submit';
