+++ src/applications/appeals/996/tests/validations/issues.unit.spec.js
@@ -3,7 +3,6 @@ import sinon from 'sinon';
 
 import { getDate } from '../../utils/dates';
 import {
-  uniqueIssue,
   maxIssues,
   missingIssueName,
   maxNameLength,
@@ -11,58 +10,6 @@ import {
 
 import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
 import { selectionRequired } from '../../../shared/validations/issues';
-
-describe('uniqueIssue', () => {
-  const _ = null;
-  const contestedIssues = [
-    {
-      attributes: {
-        ratingIssueSubjectText: 'test',
-        approxDecisionDate: '2021-01-01',
-      },
-    },
-  ];
-
-  it('should not show an error when there are no issues', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {});
-    expect(errors.addError.notCalled).to.be.true;
-  });
-  it('should not show an error when there are duplicate contested issues', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues: [contestedIssues[0], contestedIssues[0]],
-    });
-    expect(errors.addError.notCalled).to.be.true;
-  });
-  it('should not show an error when there are no duplicate issues (only date differs)', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues,
-      additionalIssues: [{ issue: 'test', decisionDate: '2021-01-02' }],
-    });
-    expect(errors.addError.notCalled).to.be.true;
-  });
-  it('should show an error when there is a duplicate additional issue', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues,
-      additionalIssues: [{ issue: 'test', decisionDate: '2021-01-01' }],
-    });
-    expect(errors.addError.called).to.be.true;
-  });
-  it('should show an error when there are multiple duplicate additional issue', () => {
-    const errors = { addError: sinon.spy() };
-    uniqueIssue(errors, _, _, _, _, _, {
-      contestedIssues,
-      additionalIssues: [
-        { issue: 'test2', decisionDate: '2021-02-01' },
-        { issue: 'test2', decisionDate: '2021-02-01' },
-      ],
-    });
-    expect(errors.addError.called).to.be.true;
-  });
-});
 
 describe('maxIssues', () => {
   it('should not show an error when the array length is less than max', () => {
