+++ src/applications/appeals/996/validations/issues.js
@@ -3,25 +3,6 @@ import { issueErrorMessages } from '../content/addIssue';
 import { maxSelectedErrorMessage } from '../../shared/content/contestableIssues';
 import { MAX_LENGTH } from '../../shared/constants';
 import { getSelected } from '../../shared/utils/issues';
-import { addUniqueIssueErrorMessage } from '../../shared/validations/issues';
-
-// Alert Veteran to duplicates based on name & decision date
-export const uniqueIssue = (
-  errors,
-  _fieldData,
-  formData,
-  _schema,
-  _uiSchema,
-  _index,
-  appStateData,
-) => {
-  addUniqueIssueErrorMessage(
-    errors,
-    formData,
-    appStateData,
-    issueErrorMessages,
-  );
-};
 
 export const maxIssues = (error, data) => {
   if (getSelected(data).length > MAX_LENGTH.SELECTIONS) {
